+++ desktop/apps/auction/templates/index.jsx
@@ -1,12 +1,19 @@
-import Layout from 'desktop/components/react/main_layout/layout.jsx'
+import { makeLayout, makePartial } from 'desktop/components/react/main_layout/layout.jsx'
 import PropTypes from 'prop-types'
 import React from 'react'
 
 export default function IndexRoute (props) {
+  const { Header, Body, Footer } = makeLayout(props)
+  const Content = makePartial(props.html)
+
   return (
-    <Layout {...props}>
-      It works!!!
-    </Layout>
+    <div>
+      <Header />
+      <Body>
+        <Content />
+      </Body>
+      <Footer />
+    </div>
   )
 }
