+++ desktop/components/react/main_layout/layout.jsx
@@ -19,3 +19,26 @@ export default function Layout ({ children, ...props }) {
 Layout.propTypes = {
   children: PropTypes.node.isRequired
 }
+
+export function makeLayout (props) {
+  return getLayout().render(props)
+}
+
+export function makePartial (html) {
+  const PartialComponent = ({ children }) => {
+    return (
+      <div>
+        <div
+          dangerouslySetInnerHTML={{ __html: html }}
+        />
+        {children}
+      </div>
+    )
+  }
+
+  PartialComponent.propTypes = {
+    children: PropTypes.node
+  }
+
+  return PartialComponent
+}
